Split HealthCertificateService responsibilities into three
