+++ src/xcode/ENA/ENA/Source/AppDelegate & Globals/AppDelegate.swift
@@ -294,6 +294,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CoronaWarnAppDelegate, Re
 			diaryStore: contactDiaryStore,
 			appConfiguration: appConfigurationProvider,
 			healthCertificateService: healthCertificateService,
+			healthCertificateRequestService: healthCertificateRequestService,
 			recycleBin: recycleBin,
 			badgeWrapper: badgeWrapper
 		)
@@ -385,10 +386,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CoronaWarnAppDelegate, Re
 		store: store,
 		dccSignatureVerifier: dccSignatureVerificationService,
 		dscListProvider: dscListProvider,
-		client: client,
 		appConfiguration: appConfigurationProvider,
 		cclService: cclService,
 		recycleBin: recycleBin
+	)
+
+	private lazy var healthCertificateRequestService = HealthCertificateRequestService(
+		store: store,
+		client: client,
+		appConfiguration: appConfigurationProvider,
+		healthCertificateService: healthCertificateService
 	)
 
 	private lazy var cclService: CCLServable = {
@@ -806,6 +813,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CoronaWarnAppDelegate, Re
 		ppacService: ppacService,
 		cclService: cclService,
 		healthCertificateService: healthCertificateService,
+		healthCertificateRequestService: healthCertificateRequestService,
 		healthCertificateValidationService: healthCertificateValidationService,
 		healthCertificateValidationOnboardedCountriesProvider: healthCertificateValidationOnboardedCountriesProvider,
 		vaccinationValueSetsProvider: vaccinationValueSetsProvider,
