+++ src/xcode/ENA/ENA/Source/Scenes/HealthCertificates/Overview/HealthCertificateOverviewViewModel.swift
@@ -11,9 +11,11 @@ class HealthCertificateOverviewViewModel {
 	// MARK: - Init
 
 	init(
-		healthCertificateService: HealthCertificateService
+		healthCertificateService: HealthCertificateService,
+		healthCertificateRequestService: HealthCertificateRequestService
 	) {
 		self.healthCertificateService = healthCertificateService
+		self.healthCertificateRequestService = healthCertificateRequestService
 
 		healthCertificateService.$healthCertifiedPersons
 			.sink {
@@ -24,7 +26,7 @@ class HealthCertificateOverviewViewModel {
 			}
 			.store(in: &subscriptions)
 
-		healthCertificateService.$testCertificateRequests
+		healthCertificateRequestService.$testCertificateRequests
 			.sink { testCertificateRequests in
 				let updatedTestCertificateRequests = testCertificateRequests
 					.sorted { $0.registrationDate > $1.registrationDate }
@@ -80,7 +82,7 @@ class HealthCertificateOverviewViewModel {
 			return
 		}
 
-		healthCertificateService.executeTestCertificateRequest(
+		healthCertificateRequestService.executeTestCertificateRequest(
 			testCertificateRequest,
 			retryIfCertificateIsPending: false
 		) { [weak self] result in
@@ -91,7 +93,7 @@ class HealthCertificateOverviewViewModel {
 	}
 
 	func remove(testCertificateRequest: TestCertificateRequest) {
-		healthCertificateService.remove(testCertificateRequest: testCertificateRequest)
+		healthCertificateRequestService.remove(testCertificateRequest: testCertificateRequest)
 	}
 
 	func attemptToRestoreDecodingFailedHealthCertificates() {
@@ -101,6 +103,7 @@ class HealthCertificateOverviewViewModel {
 	// MARK: - Private
 
 	private let healthCertificateService: HealthCertificateService
+	private let healthCertificateRequestService: HealthCertificateRequestService
 	private var subscriptions = Set<AnyCancellable>()
 
 }
