+++ src/xcode/ENA/ENA/Source/Services/CoronaTestService/CoronaTestService.swift
@@ -25,6 +25,7 @@ class CoronaTestService {
 		diaryStore: DiaryStoring,
 		appConfiguration: AppConfigurationProviding,
 		healthCertificateService: HealthCertificateService,
+		healthCertificateRequestService: HealthCertificateRequestService,
 		notificationCenter: UserNotificationCenter = UNUserNotificationCenter.current(),
 		recycleBin: RecycleBin,
 		badgeWrapper: HomeBadgeWrapper
@@ -43,6 +44,7 @@ class CoronaTestService {
 			self.appConfiguration = CachedAppConfigurationMock()
 
 			self.healthCertificateService = healthCertificateService
+			self.healthCertificateRequestService = healthCertificateRequestService
 			self.notificationCenter = notificationCenter
 			self.recycleBin = recycleBin
 			self.badgeWrapper = badgeWrapper
@@ -66,6 +68,7 @@ class CoronaTestService {
 		self.diaryStore = diaryStore
 		self.appConfiguration = appConfiguration
 		self.healthCertificateService = healthCertificateService
+		self.healthCertificateRequestService = healthCertificateRequestService
 		self.notificationCenter = notificationCenter
 		self.recycleBin = recycleBin
 		self.badgeWrapper = badgeWrapper
@@ -73,7 +76,7 @@ class CoronaTestService {
 		self.fakeRequestService = FakeRequestService(client: client, restServiceProvider: restServiceProvider)
 		self.warnOthersReminder = WarnOthersReminder(store: store)
 
-		healthCertificateService.didRegisterTestCertificate = setUniqueCertificateIdentifier
+		healthCertificateRequestService.didRegisterTestCertificate = setUniqueCertificateIdentifier
 
 		setup()
 	}
@@ -87,6 +90,7 @@ class CoronaTestService {
 		diaryStore: DiaryStoring,
 		appConfiguration: AppConfigurationProviding,
 		healthCertificateService: HealthCertificateService,
+		healthCertificateRequestService: HealthCertificateRequestService,
 		notificationCenter: UserNotificationCenter = UNUserNotificationCenter.current(),
 		recycleBin: RecycleBin,
 		badgeWrapper: HomeBadgeWrapper
@@ -99,6 +103,7 @@ class CoronaTestService {
 			diaryStore: diaryStore,
 			appConfiguration: appConfiguration,
 			healthCertificateService: healthCertificateService,
+			healthCertificateRequestService: healthCertificateRequestService,
 			notificationCenter: notificationCenter,
 			recycleBin: recycleBin,
 			badgeWrapper: badgeWrapper
@@ -667,6 +672,7 @@ class CoronaTestService {
 	private let diaryStore: DiaryStoring
 	private let appConfiguration: AppConfigurationProviding
 	private let healthCertificateService: HealthCertificateService
+	private let healthCertificateRequestService: HealthCertificateRequestService
 	private let notificationCenter: UserNotificationCenter
 	private let recycleBin: RecycleBin
 	private let badgeWrapper: HomeBadgeWrapper
@@ -916,7 +922,7 @@ class CoronaTestService {
 						}
 
 						if testResult == .negative && coronaTest.certificateConsentGiven && !coronaTest.certificateRequested {
-							self.healthCertificateService.registerAndExecuteTestCertificateRequest(
+							self.healthCertificateRequestService.registerAndExecuteTestCertificateRequest(
 								coronaTestType: coronaTestType,
 								registrationToken: registrationToken,
 								registrationDate: registrationDate,
