+++ src/xcode/ENA/ENA/Source/Services/HealthCertificate/HealthCertificateService.swift
@@ -5,7 +5,6 @@
 import UIKit
 import OpenCombine
 import HealthCertificateToolkit
-import UserNotifications
 
 // global to access in unit tests
 // version will be used for migration logic
@@ -20,7 +19,6 @@ class HealthCertificateService {
 		store: HealthCertificateStoring,
 		dccSignatureVerifier: DCCSignatureVerifying,
 		dscListProvider: DSCListProviding,
-		client: Client,
 		appConfiguration: AppConfigurationProviding,
 		digitalCovidCertificateAccess: DigitalCovidCertificateAccessProtocol = DigitalCovidCertificateAccess(),
 		notificationCenter: UserNotificationCenter = UNUserNotificationCenter.current(),
@@ -30,16 +28,20 @@ class HealthCertificateService {
 		#if DEBUG
 		if isUITesting {
 			let store = MockTestStore()
+			let appConfiguration = CachedAppConfigurationMock(store: store)
 
 			self.store = store
 			self.dccSignatureVerifier = dccSignatureVerifier
 			self.dscListProvider = DSCListProvider(client: CachingHTTPClientMock(), store: store)
-			self.client = ClientMock()
-			self.appConfiguration = CachedAppConfigurationMock(store: store)
+			self.appConfiguration = appConfiguration
 			self.digitalCovidCertificateAccess = digitalCovidCertificateAccess
-			self.notificationCenter = notificationCenter
+			self.healthCertificateNotificationService = HealthCertificateNotificationService(
+				appConfiguration: appConfiguration,
+				notificationCenter: notificationCenter
+			)
 			self.cclService = cclService
 			self.recycleBin = recycleBin
+
 			setup()
 			configureForTesting()
 
@@ -50,10 +52,12 @@ class HealthCertificateService {
 		self.store = store
 		self.dccSignatureVerifier = dccSignatureVerifier
 		self.dscListProvider = dscListProvider
-		self.client = client
 		self.appConfiguration = appConfiguration
 		self.digitalCovidCertificateAccess = digitalCovidCertificateAccess
-		self.notificationCenter = notificationCenter
+		self.healthCertificateNotificationService = HealthCertificateNotificationService(
+			appConfiguration: appConfiguration,
+			notificationCenter: notificationCenter
+		)
 		self.cclService = cclService
 		self.recycleBin = recycleBin
 
@@ -83,20 +87,7 @@ class HealthCertificateService {
 		}
 	}
 
-	@DidSetPublished private(set) var testCertificateRequests = [TestCertificateRequest]() {
-		didSet {
-			Log.debug("Did set testCertificateRequests.")
-
-			if initialTestCertificateRequestsReadFromStore {
-				store.testCertificateRequests = testCertificateRequests
-			}
-
-			updateTestCertificateRequestSubscriptions(for: testCertificateRequests)
-		}
-	}
-
 	private(set) var unseenNewsCount = CurrentValueSubject<Int, Never>(0)
-	var didRegisterTestCertificate: ((String, TestCertificateRequest) -> Void)?
 	
 	var nextValidityTimer: Timer?
 
@@ -246,7 +237,7 @@ class HealthCertificateService {
 		scheduleTimer()
 
 		if healthCertificate.type != .test {
-			createNotifications(for: healthCertificate)
+			healthCertificateNotificationService.createNotifications(for: healthCertificate)
 		}
 
 		if isNewPerson {
@@ -276,7 +267,7 @@ class HealthCertificateService {
 			}
 		}
 		// we do not have to wait here, so we leave the completion empty
-		removeAllNotifications(for: healthCertificate, completion: {})
+		healthCertificateNotificationService.removeAllNotifications(for: healthCertificate, completion: {})
 
 		// Move HealthCertificate to the recycle-bin
 		recycleBin.moveToBin(.certificate(healthCertificate))
@@ -303,171 +294,11 @@ class HealthCertificateService {
 		}
 	}
 
-	func registerAndExecuteTestCertificateRequest(
-		coronaTestType: CoronaTestType,
-		registrationToken: String,
-		registrationDate: Date,
-		retryExecutionIfCertificateIsPending: Bool,
-		labId: String?,
-		completion: ((Result<Void, HealthCertificateServiceError.TestCertificateRequestError>) -> Void)? = nil
-	) {
-		Log.info("[HealthCertificateService] Registering test certificate request: (coronaTestType: \(coronaTestType), registrationToken: \(private: registrationToken), registrationDate: \(registrationDate), retryExecutionIfCertificateIsPending: \(retryExecutionIfCertificateIsPending)", log: .api)
-
-		let testCertificateRequest = TestCertificateRequest(
-			coronaTestType: coronaTestType,
-			registrationToken: registrationToken,
-			registrationDate: registrationDate,
-			labId: labId
-		)
-
-		testCertificateRequests.append(testCertificateRequest)
-
-		executeTestCertificateRequest(
-			testCertificateRequest,
-			retryIfCertificateIsPending: retryExecutionIfCertificateIsPending,
-			completion: completion
-		)
-	}
-
-	// swiftlint:disable:next cyclomatic_complexity
-	func executeTestCertificateRequest(
-		_ testCertificateRequest: TestCertificateRequest,
-		retryIfCertificateIsPending: Bool,
-		completion: ((Result<Void, HealthCertificateServiceError.TestCertificateRequestError>) -> Void)? = nil
-	) {
-		Log.info("[HealthCertificateService] Executing test certificate request: \(private: testCertificateRequest)", log: .api)
-
-		testCertificateRequest.isLoading = true
-
-		// If we didn't retrieve a labId for a PRC test result, the lab is not supporting test certificates.
-		if testCertificateRequest.coronaTestType == .pcr && testCertificateRequest.labId == nil {
-			testCertificateRequest.requestExecutionFailed = true
-			testCertificateRequest.isLoading = false
-			completion?(.failure(.dgcNotSupportedByLab))
-			return
-		}
-
-		do {
-			let rsaKeyPair = try testCertificateRequest.rsaKeyPair ?? DCCRSAKeyPair(registrationToken: testCertificateRequest.registrationToken)
-			testCertificateRequest.rsaKeyPair = rsaKeyPair
-			let publicKey = try rsaKeyPair.publicKeyForBackend()
-
-			appConfiguration.appConfiguration()
-				.sink { [weak self] in
-					guard let self = self else { return }
-
-					var waitAfterPublicKeyRegistrationInSeconds = TimeInterval($0.dgcParameters.testCertificateParameters.waitAfterPublicKeyRegistrationInSeconds)
-
-					var waitForRetryInSeconds = TimeInterval($0.dgcParameters.testCertificateParameters.waitForRetryInSeconds)
-
-					// 0 means the value is not set -> setting it to a default waiting time of 10 seconds
-					if waitAfterPublicKeyRegistrationInSeconds == 0 {
-						waitAfterPublicKeyRegistrationInSeconds = 10
-					}
-
-					if waitForRetryInSeconds == 0 {
-						waitForRetryInSeconds = 10
-					}
-
-					Log.info("[HealthCertificateService] waitAfterPublicKeyRegistrationInSeconds: \(waitAfterPublicKeyRegistrationInSeconds), waitForRetryInSeconds: \(waitForRetryInSeconds)", log: .api)
-
-					if !testCertificateRequest.rsaPublicKeyRegistered {
-						Log.info("[HealthCertificateService] Registering public key …", log: .api)
-
-						self.client.dccRegisterPublicKey(
-							isFake: false,
-							token: testCertificateRequest.registrationToken,
-							publicKey: publicKey,
-							completion: { result in
-								switch result {
-								case .success:
-									Log.info("[HealthCertificateService] Public key successfully registered", log: .api)
-
-									testCertificateRequest.rsaPublicKeyRegistered = true
-									DispatchQueue.global().asyncAfter(deadline: .now() + waitAfterPublicKeyRegistrationInSeconds) {
-										self.requestDigitalCovidCertificate(
-											for: testCertificateRequest,
-											rsaKeyPair: rsaKeyPair,
-											retryIfCertificateIsPending: retryIfCertificateIsPending,
-											waitForRetryInSeconds: waitForRetryInSeconds,
-											completion: completion
-										)
-									}
-								case .failure(let registrationError) where registrationError == .tokenAlreadyAssigned:
-									Log.info("[HealthCertificateService] Public key was already registered.", log: .api)
-
-									testCertificateRequest.rsaPublicKeyRegistered = true
-									testCertificateRequest.isLoading = false
-									self.requestDigitalCovidCertificate(
-										for: testCertificateRequest,
-										rsaKeyPair: rsaKeyPair,
-										retryIfCertificateIsPending: retryIfCertificateIsPending,
-										waitForRetryInSeconds: waitForRetryInSeconds,
-										completion: completion
-									)
-								case .failure(let registrationError):
-									Log.error("[HealthCertificateService] Public key registration failed: \(registrationError.localizedDescription)", log: .api)
-
-									testCertificateRequest.requestExecutionFailed = true
-									testCertificateRequest.isLoading = false
-									completion?(.failure(.publicKeyRegistrationFailed(registrationError)))
-								}
-							}
-						)
-					} else if let encryptedDEK = testCertificateRequest.encryptedDEK,
-							  let encryptedCOSE = testCertificateRequest.encryptedCOSE {
-						Log.info("[HealthCertificateService] Encrypted COSE and DEK already exist, immediately assembling certificate.", log: .api)
-
-						self.assembleDigitalCovidCertificate(
-							for: testCertificateRequest,
-							rsaKeyPair: rsaKeyPair,
-							encryptedDEK: encryptedDEK,
-							encryptedCOSE: encryptedCOSE,
-							completion: completion
-						)
-					} else {
-						Log.info("[HealthCertificateService] Public key already registered, immediately requesting certificate.", log: .api)
-
-						self.requestDigitalCovidCertificate(
-							for: testCertificateRequest,
-							rsaKeyPair: rsaKeyPair,
-							retryIfCertificateIsPending: retryIfCertificateIsPending,
-							waitForRetryInSeconds: waitForRetryInSeconds,
-							completion: completion
-						)
-					}
-				}
-				.store(in: &subscriptions)
-		} catch let error as DCCRSAKeyPairError {
-			Log.error("[HealthCertificateService] Key pair error occurred: \(error.localizedDescription)", log: .api)
-
-			testCertificateRequest.requestExecutionFailed = true
-			testCertificateRequest.isLoading = false
-			completion?(.failure(.rsaKeyPairGenerationFailed(error)))
-		} catch {
-			Log.error("[HealthCertificateService] Error occurred: \(error.localizedDescription)", log: .api)
-
-			testCertificateRequest.requestExecutionFailed = true
-			testCertificateRequest.isLoading = false
-			completion?(.failure(.other(error)))
-		}
-	}
-
-	func remove(testCertificateRequest: TestCertificateRequest) {
-		testCertificateRequest.rsaKeyPair?.removeFromKeychain()
-		if let index = testCertificateRequests.firstIndex(of: testCertificateRequest) {
-			testCertificateRequests.remove(at: index)
-		}
-	}
-
 	func updatePublishersFromStore() {
 		Log.info("[HealthCertificateService] Updating publishers from store", log: .api)
 
 		healthCertifiedPersons = store.healthCertifiedPersons
 		initialHealthCertifiedPersonsReadFromStore = true
-
-		testCertificateRequests = store.testCertificateRequests
-		initialTestCertificateRequestsReadFromStore = true
 
 		updateHealthCertifiedPersonSubscriptions(for: healthCertifiedPersons)
 	}
@@ -541,7 +372,7 @@ class HealthCertificateService {
 		healthCertifiedPersons.forEach { healthCertifiedPerson in
 			healthCertifiedPerson.healthCertificates.forEach { healthCertificate in
 				updateValidityState(for: healthCertificate)
-				updateNotifications(for: healthCertificate)
+				healthCertificateNotificationService.recreateNotifications(for: healthCertificate)
 			}
 		}
 
@@ -620,18 +451,15 @@ class HealthCertificateService {
 	private let store: HealthCertificateStoring
 	private let dccSignatureVerifier: DCCSignatureVerifying
 	private let dscListProvider: DSCListProviding
-	private let client: Client
 	private let appConfiguration: AppConfigurationProviding
 	private let digitalCovidCertificateAccess: DigitalCovidCertificateAccessProtocol
-	private let notificationCenter: UserNotificationCenter
+	private let healthCertificateNotificationService: HealthCertificateNotificationService
 	private let recycleBin: RecycleBin
 	private let cclService: CCLServable
 
 	private var initialHealthCertifiedPersonsReadFromStore = false
-	private var initialTestCertificateRequestsReadFromStore = false
 
 	private var healthCertifiedPersonSubscriptions = Set<AnyCancellable>()
-	private var testCertificateRequestSubscriptions = Set<AnyCancellable>()
 	private var subscriptions = Set<AnyCancellable>()
 
 	private func setup() {
@@ -639,7 +467,6 @@ class HealthCertificateService {
 		updatePublishersFromStore()
 		updateTimeBasedValidityStates()
 
-		subscribeToNotifications()
 		updateGradients()
 		
 		subscribeAppConfigUpdates()
@@ -743,7 +570,7 @@ class HealthCertificateService {
 				}
 				#endif
 
-				self.scheduleBoosterNotificationIfNeeded(
+				self.healthCertificateNotificationService.scheduleBoosterNotificationIfNeeded(
 					for: person,
 					previousBoosterNotificationIdentifier: previousBoosterNotificationIdentifier,
 					completion: completion
@@ -812,374 +639,6 @@ class HealthCertificateService {
 		} else {
 			healthCertificate.validityState = .valid
 		}
-	}
-
-	private func updateNotifications(for healthCertificate: HealthCertificate) {
-		// No notifications for test certificates
-		if healthCertificate.type == .recovery || healthCertificate.type == .vaccination {
-			removeAllNotifications(for: healthCertificate, completion: { [weak self] in
-				self?.createNotifications(for: healthCertificate)
-			})
-		}
-	}
-
-	private func updateTestCertificateRequestSubscriptions(for testCertificateRequests: [TestCertificateRequest]) {
-		Log.debug("Update test certificate subscriptions.")
-
-		testCertificateRequestSubscriptions = []
-
-		testCertificateRequests.forEach { testCertificateRequest in
-			testCertificateRequest.objectDidChange
-				.sink { [weak self] _ in
-					guard let self = self else { return }
-					// Trigger publisher to inform subscribers and update store
-					self.testCertificateRequests = self.testCertificateRequests
-				}
-				.store(in: &testCertificateRequestSubscriptions)
-		}
-	}
-
-	private func subscribeToNotifications() {
-		NotificationCenter.default.ocombine
-			.publisher(for: UIApplication.didBecomeActiveNotification)
-			.sink { [weak self] _ in
-				self?.testCertificateRequests.forEach {
-					self?.executeTestCertificateRequest($0, retryIfCertificateIsPending: false)
-				}
-			}
-			.store(in: &subscriptions)
-	}
-
-	private func requestDigitalCovidCertificate(
-		for testCertificateRequest: TestCertificateRequest,
-		rsaKeyPair: DCCRSAKeyPair,
-		retryIfCertificateIsPending: Bool,
-		waitForRetryInSeconds: TimeInterval,
-		completion: ((Result<Void, HealthCertificateServiceError.TestCertificateRequestError>) -> Void)?
-	) {
-		Log.info("[HealthCertificateService] Requesting certificate…", log: .api)
-
-		client.getDigitalCovid19Certificate(
-			registrationToken: testCertificateRequest.registrationToken,
-			isFake: false
-		) { [weak self] result in
-			switch result {
-			case .success(let dccResponse):
-				Log.info("[HealthCertificateService] Certificate request succeeded", log: .api)
-
-				self?.assembleDigitalCovidCertificate(
-					for: testCertificateRequest,
-					rsaKeyPair: rsaKeyPair,
-					encryptedDEK: dccResponse.dek,
-					encryptedCOSE: dccResponse.dcc,
-					completion: completion
-				)
-			case .failure(let error) where error == .dccPending && retryIfCertificateIsPending:
-				DispatchQueue.global().asyncAfter(deadline: .now() + waitForRetryInSeconds) {
-					Log.info("[HealthCertificateService] Certificate request failed with .dccPending, retrying.", log: .api)
-
-					self?.requestDigitalCovidCertificate(
-						for: testCertificateRequest,
-						rsaKeyPair: rsaKeyPair,
-						retryIfCertificateIsPending: false,
-						waitForRetryInSeconds: waitForRetryInSeconds,
-						completion: completion
-					)
-				}
-			case .failure(let error):
-				Log.error("[HealthCertificateService] Certificate request failed with error \(error.localizedDescription)", log: .api)
-
-				testCertificateRequest.requestExecutionFailed = true
-				testCertificateRequest.isLoading = false
-				completion?(.failure(.certificateRequestFailed(error)))
-			}
-		}
-	}
-
-	private func assembleDigitalCovidCertificate(
-		for testCertificateRequest: TestCertificateRequest,
-		rsaKeyPair: DCCRSAKeyPair,
-		encryptedDEK: String,
-		encryptedCOSE: String,
-		completion: ((Result<Void, HealthCertificateServiceError.TestCertificateRequestError>) -> Void)?
-	) {
-		Log.info("[HealthCertificateService] Assembling certificate…", log: .api)
-
-		guard let encryptedDEKData = Data(base64Encoded: encryptedDEK) else {
-			Log.error("[HealthCertificateService] Assembling certificate failed: base64 decoding failed", log: .api)
-
-			testCertificateRequest.requestExecutionFailed = true
-			testCertificateRequest.isLoading = false
-			completion?(.failure(.base64DecodingFailed))
-			return
-		}
-
-		do {
-			let decodedDEK = try rsaKeyPair.decrypt(encryptedDEKData)
-			let result = digitalCovidCertificateAccess.convertToBase45(from: encryptedCOSE, with: decodedDEK)
-
-			switch result {
-			case .success(let healthCertificateBase45):
-				let registerResult = registerHealthCertificate(
-					base45: healthCertificateBase45,
-					checkSignatureUpfront: false,
-					checkMaxPersonCount: false,
-					markAsNew: true
-				)
-
-				switch registerResult {
-				case .success(let certificateResult):
-					Log.info("[HealthCertificateService] Certificate assembly succeeded", log: .api)
-					
-					didRegisterTestCertificate?(certificateResult.certificate.uniqueCertificateIdentifier, testCertificateRequest)
-					
-					remove(testCertificateRequest: testCertificateRequest)
-					completion?(.success(()))
-				case .failure(let error):
-					Log.error("[HealthCertificateService] Assembling certificate failed: Register failed: \(error.localizedDescription)", log: .api)
-
-					testCertificateRequest.requestExecutionFailed = true
-					testCertificateRequest.isLoading = false
-					completion?(.failure(.registrationError(error)))
-				}
-			case .failure(let error):
-				Log.error("[HealthCertificateService] Assembling certificate failed: Conversion failed: \(error.localizedDescription)", log: .api)
-
-				testCertificateRequest.requestExecutionFailed = true
-				testCertificateRequest.isLoading = false
-				completion?(.failure(.assemblyFailed(error)))
-			}
-		} catch {
-			Log.error("[HealthCertificateService] Assembling certificate failed: DEK decryption failed: \(error.localizedDescription)", log: .api)
-
-			testCertificateRequest.requestExecutionFailed = true
-			testCertificateRequest.isLoading = false
-			completion?(.failure(.decryptionFailed(error)))
-		}
-	}
-	
-	private func removeAllNotifications(
-		for healthCertificate: HealthCertificate,
-		completion: @escaping () -> Void
-	) {
-		let id = healthCertificate.uniqueCertificateIdentifier
-		
-		Log.info("Cancel all notifications for certificate with id: \(private: id).", log: .vaccination)
-		
-		let expiringSoonId = LocalNotificationIdentifier.certificateExpiringSoon.rawValue + "\(id)"
-		let expiredId = LocalNotificationIdentifier.certificateExpired.rawValue + "\(id)"
-
-		notificationCenter.getPendingNotificationRequests { [weak self] requests in
-			let notificationIds = requests.map {
-				$0.identifier
-			}.filter {
-				$0.contains(expiringSoonId) ||
-				$0.contains(expiredId)
-			}
-
-			self?.notificationCenter.removePendingNotificationRequests(withIdentifiers: notificationIds)
-			completion()
-		}
-	}
-	
-	private func createNotifications(for healthCertificate: HealthCertificate) {
-		Log.info("Create notifications.")
-
-		let id = healthCertificate.uniqueCertificateIdentifier
-		
-		let expirationThresholdInDays = appConfiguration.currentAppConfig.value.dgcParameters.expirationThresholdInDays
-		let expiringSoonDate = Calendar.current.date(
-			byAdding: .day,
-			value: -Int(expirationThresholdInDays),
-			to: healthCertificate.expirationDate
-		)
-		
-		let expirationDate = healthCertificate.expirationDate
-		scheduleNotificationForExpiringSoon(id: id, date: expiringSoonDate)
-		scheduleNotificationForExpired(id: id, date: expirationDate)
-
-		// Schedule an 'invalid' notification, if it was not scheduled before.
-		if healthCertificate.validityState == .invalid && !healthCertificate.didShowInvalidNotification {
-			scheduleInvalidNotification(id: id)
-			healthCertificate.didShowInvalidNotification = true
-		}
-
-		// Schedule a 'blocked' notification, if it was not scheduled before.
-		if healthCertificate.validityState == .blocked && !healthCertificate.didShowBlockedNotification {
-			scheduleBlockedNotification(id: id)
-			healthCertificate.didShowBlockedNotification = true
-		}
-	}
-	
-	private func scheduleNotificationForExpiringSoon(
-		id: String,
-		date: Date?
-	) {
-		guard let date = date else {
-			Log.error("Could not schedule expiring soon notification for certificate with id: \(private: id) because we have no expiringSoonDate.", log: .vaccination)
-			return
-		}
-		
-		Log.info("Schedule expiring soon notification for certificate with id: \(private: id) with expiringSoonDate: \(date)", log: .vaccination)
-
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.LocalNotifications.certificateGenericTitle
-		content.body = AppStrings.LocalNotifications.certificateValidityBody
-		content.sound = .default
-
-		let expiringSoonDateComponents = Calendar.current.dateComponents(
-			[.year, .month, .day, .hour, .minute, .second],
-			from: date
-		)
-
-		let trigger = UNCalendarNotificationTrigger(dateMatching: expiringSoonDateComponents, repeats: false)
-
-		let request = UNNotificationRequest(
-			identifier: LocalNotificationIdentifier.certificateExpiringSoon.rawValue + "\(id)",
-			content: content,
-			trigger: trigger
-		)
-
-		addNotification(request: request)
-	}
-	
-	private func scheduleNotificationForExpired(
-		id: String,
-		date: Date
-	) {
-		Log.info("Schedule expired notification for certificate with id: \(private: id) with expirationDate: \(date)", log: .vaccination)
-
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.LocalNotifications.certificateGenericTitle
-		content.body = AppStrings.LocalNotifications.certificateValidityBody
-		content.sound = .default
-
-		let expiredDateComponents = Calendar.current.dateComponents(
-			[.year, .month, .day, .hour, .minute, .second],
-			from: date
-		)
-
-		let trigger = UNCalendarNotificationTrigger(dateMatching: expiredDateComponents, repeats: false)
-
-		let request = UNNotificationRequest(
-			identifier: LocalNotificationIdentifier.certificateExpired.rawValue + "\(id)",
-			content: content,
-			trigger: trigger
-		)
-
-		addNotification(request: request)
-	}
-
-	private func scheduleInvalidNotification(
-		id: String
-	) {
-		Log.info("Schedule invalid notification for certificate with id: \(private: id)", log: .vaccination)
-
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.LocalNotifications.certificateGenericTitle
-		content.body = AppStrings.LocalNotifications.certificateValidityBody
-		content.sound = .default
-
-		let request = UNNotificationRequest(
-			identifier: LocalNotificationIdentifier.certificateInvalid.rawValue + "\(id)",
-			content: content,
-			trigger: nil
-		)
-
-		addNotification(request: request)
-	}
-
-	private func scheduleBlockedNotification(
-		id: String
-	) {
-		Log.info("Schedule blocked notification for certificate with id: \(private: id)", log: .vaccination)
-
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.LocalNotifications.certificateGenericTitle
-		content.body = AppStrings.LocalNotifications.certificateValidityBody
-		content.sound = .default
-
-		let request = UNNotificationRequest(
-			identifier: LocalNotificationIdentifier.certificateBlocked.rawValue + "\(id)",
-			content: content,
-			trigger: nil
-		)
-
-		addNotification(request: request)
-	}
-	
-	private func addNotification(request: UNNotificationRequest, completion: (() -> Void)? = nil) {
-		_ = notificationCenter.getPendingNotificationRequests { [weak self] requests in
-			guard !requests.contains(request) else {
-				Log.info(
-					"Did not schedule notification: \(private: request.identifier) because it is already scheduled.",
-					log: .vaccination
-				)
-				completion?()
-
-				return
-			}
-			self?.notificationCenter.add(request) { error in
-				if error != nil {
-					Log.error(
-						"Could not schedule notification: \(private: request.identifier)",
-						log: .vaccination,
-						error: error
-					)
-				}
-
-				completion?()
-			}
-		}
-	}
-	
-	private func scheduleBoosterNotificationIfNeeded(
-		for person: HealthCertifiedPerson,
-		previousBoosterNotificationIdentifier: String?,
-		completion: (() -> Void)? = nil
-	) {
-		let name = person.name?.standardizedName
-		guard let newBoosterNotificationIdentifier = person.dccWalletInfo?.boosterNotification.identifier else {
-			Log.info("No booster notification identifier found for person \(private: String(describing: name))", log: .vaccination)
-			completion?()
-
-			return
-		}
-
-		if newBoosterNotificationIdentifier != previousBoosterNotificationIdentifier {
-			// we need to have an ID for the notification and since the certified person doesn't have this property "unlike the certificates" we will compute it as the hash of the string of the standardizedName + dateOfBirth
-			guard let name = name, let dateOfBirth = person.dateOfBirth else {
-				Log.error("standardizedName or dateOfBirth is nil, will not trigger booster notification", log: .vaccination)
-				completion?()
-
-				return
-			}
-
-			Log.info("Scheduling booster notification for \(private: String(describing: name))", log: .vaccination)
-
-			let id = ENAHasher.sha256(name + dateOfBirth)
-			self.scheduleBoosterNotification(id: id, completion: completion)
-		} else {
-			Log.debug("Booster notification identifier \(private: newBoosterNotificationIdentifier) unchanged, no booster notification scheduled", log: .vaccination)
-			completion?()
-		}
-	}
-	
-	private func scheduleBoosterNotification(id: String, completion: (() -> Void)? = nil) {
-		Log.info("Schedule booster notification for certificate with id: \(private: id) with trigger date: \(Date())", log: .vaccination)
-
-		let content = UNMutableNotificationContent()
-		content.title = AppStrings.LocalNotifications.certificateGenericTitle
-		content.body = AppStrings.LocalNotifications.certificateGenericBody
-		content.sound = .default
-
-		let request = UNNotificationRequest(
-			identifier: LocalNotificationIdentifier.boosterVaccination.rawValue + "\(id)",
-			content: content,
-			trigger: nil
-		)
-
-		addNotification(request: request, completion: completion)
 	}
 
 	// swiftlint:disable:next file_length
